Pull out useChatOptions into own file in barrel
