+++ packages/app/ui/contexts/chatOptions/index.ts
@@ -1,1 +1,2 @@
 export * from './chatOptions';
+export * from './useChatOptions';
